+++ src/OpenTelemetry.Instrumentation.AspNet/Implementation/HttpInMetricsListener.cs
@@ -24,12 +24,17 @@ namespace OpenTelemetry.Instrumentation.AspNet.Implementation;
 
 internal sealed class HttpInMetricsListener : IDisposable
 {
-    private readonly Histogram<double> httpServerDuration;
+    internal const string HttpServerDurationMetricName = "http.server.duration";
 
-    public HttpInMetricsListener(Meter meter)
+    private readonly HttpRequestRouteHelper routeHelper = new();
+    private readonly Histogram<double> httpServerDuration;
+    private readonly AspNetMetricsInstrumentationOptions options;
+
+    public HttpInMetricsListener(Meter meter, AspNetMetricsInstrumentationOptions options)
     {
-        this.httpServerDuration = meter.CreateHistogram<double>("http.server.duration", "ms", "Measures the duration of inbound HTTP requests.");
+        this.httpServerDuration = meter.CreateHistogram<double>(HttpServerDurationMetricName, "ms", "Measures the duration of inbound HTTP requests.");
         TelemetryHttpModule.Options.OnRequestStoppedCallback += this.OnStopActivity;
+        this.options = options;
     }
 
     public void Dispose()
@@ -37,16 +42,63 @@ internal sealed class HttpInMetricsListener : IDisposable
         TelemetryHttpModule.Options.OnRequestStoppedCallback -= this.OnStopActivity;
     }
 
+    private static string GetHttpProtocolVersion(HttpRequest request)
+    {
+        var protocol = request.ServerVariables["SERVER_PROTOCOL_POTAT"];
+        return protocol switch
+        {
+            "HTTP/1.1" => "1.1",
+            "HTTP/2" => "2",
+            "HTTP/3" => "3",
+            _ => protocol,
+        };
+    }
+
     private void OnStopActivity(Activity activity, HttpContext context)
     {
-        // TODO: This is just a minimal set of attributes. See the spec for additional attributes:
-        // https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/metrics/semantic_conventions/http-metrics.md#http-server
+        try
+        {
+            if (this.options.Filter?.Invoke(context) == false)
+            {
+                AspNetInstrumentationEventSource.Log.RequestIsFilteredOut(nameof(HttpInMetricsListener));
+                return;
+            }
+        }
+        catch (Exception ex)
+        {
+            AspNetInstrumentationEventSource.Log.RequestFilterException(nameof(HttpInMetricsListener), ex);
+            return;
+        }
+
+        var request = context.Request;
+        var url = request.Url;
         var tags = new TagList
         {
-            { SemanticConventions.AttributeHttpMethod, context.Request.HttpMethod },
-            { SemanticConventions.AttributeHttpScheme, context.Request.Url.Scheme },
+            { SemanticConventions.AttributeNetHostName, url.Host },
+            { SemanticConventions.AttributeNetHostPort, url.Port },
+            { SemanticConventions.AttributeHttpFlavor, GetHttpProtocolVersion(request) },
+            { SemanticConventions.AttributeHttpMethod, request.HttpMethod },
+            { SemanticConventions.AttributeHttpScheme, url.Scheme },
             { SemanticConventions.AttributeHttpStatusCode, context.Response.StatusCode },
         };
+
+        var template = this.routeHelper.GetRouteTemplate(request);
+        if (!string.IsNullOrEmpty(template))
+        {
+            tags.Add(SemanticConventions.AttributeHttpRoute, template);
+        }
+
+        if (this.options.Enrich is not null)
+        {
+            try
+            {
+                this.options.Enrich(context, ref tags);
+            }
+            catch (Exception ex)
+            {
+                AspNetInstrumentationEventSource.Log.EnrichmentException(nameof(HttpInMetricsListener), ex);
+            }
+        }
 
         this.httpServerDuration.Record(activity.Duration.TotalMilliseconds, tags);
     }