+++ test/OpenTelemetry.Instrumentation.AspNet.Tests/HttpInMetricsListenerTests.cs
@@ -16,6 +16,8 @@
 
 using System.Collections.Generic;
 using System.IO;
+using System.Reflection;
+using System.Threading;
 using System.Web;
 using OpenTelemetry.Context.Propagation;
 using OpenTelemetry.Metrics;
@@ -56,7 +58,10 @@ public class HttpInMetricsListenerTests
             .Build();
 
         var activity = ActivityHelper.StartAspNetActivity(Propagators.DefaultTextMapPropagator, HttpContext.Current, TelemetryHttpModule.Options.OnRequestStartedCallback);
+        Thread.Sleep(1); // Make sure duration is always greater than 0 to avoid flakiness.
         ActivityHelper.StopAspNetActivity(Propagators.DefaultTextMapPropagator, activity, HttpContext.Current, TelemetryHttpModule.Options.OnRequestStoppedCallback);
+
+        Assert.True(duration > 0, "Metruc duration should be set.");
 
         meterProvider.ForceFlush();
 
@@ -64,7 +69,7 @@ public class HttpInMetricsListenerTests
         foreach (var p in exportedItems[0].GetMetricPoints())
         {
             metricPoints.Add(p);
-        }
+        } 
 
         Assert.Single(metricPoints);