+++ test/OpenTelemetry.Instrumentation.AspNet.Tests/HttpInMetricsListenerTests.cs
@@ -60,8 +60,6 @@ public class HttpInMetricsListenerTests
         Thread.Sleep(1); // Make sure duration is always greater than 0 to avoid flakiness.
         ActivityHelper.StopAspNetActivity(Propagators.DefaultTextMapPropagator, activity, HttpContext.Current, TelemetryHttpModule.Options.OnRequestStoppedCallback);
 
-        Assert.True(duration > 0, "Metric duration should be set.");
-
         meterProvider.ForceFlush();
 
         var metricPoints = new List<MetricPoint>();
@@ -81,6 +79,7 @@ public class HttpInMetricsListenerTests
         Assert.Equal("http.server.duration", exportedItems[0].Name);
         Assert.Equal(1L, count);
         Assert.Equal(duration, sum);
+        Assert.True(duration > 0, "Metric duration should be set.");
 
         Assert.Equal(3, metricPoints[0].Tags.Count);
         string? httpMethod = null;