+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## Unreleased
+
 ## 1.6.0-beta.1
 
 Released 2023-Oct-11