+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -2,13 +2,15 @@
 
 ## Unreleased
 
-* Added enrich and filter functionality to metric instrumentation [#TODO].
+* Added enrich and filter functionality to metric instrumentation
+  ([#1407](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1407)).
 
   * New overload of `AddAspnetInstrumentation` now accepts a configuration delegate.
   * The `Filter` delegate is used to decide if a metric should be collected.
   * The `Enrich` can be used to add additional metric tags.
 
-* Additional metric tags will now be emitted [#TODO].
+* Additional metric tags will now be emitted
+  ([#1407](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1407)).
 
 ## 1.6.0-beta.1