+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fix for `http.server.duration` being time always being zero since
+  ([#1388](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1388)).
+
 ## 1.6.0-beta.1
 
 Released 2023-Oct-11