+++ src/OpenTelemetry.Instrumentation.AspNet/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 1.6.0-beta.2
+
+Released 2023-Nov-06
+
 * Fixed an issue that caused `http.server.duration` metric value to always be set
   to `0`. The issue exists in `1.6.0-beta.1` version.
   ([#1425](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1425))